+++ Source/Services/Mangos.World/Player/WS_Player_Initializator.cs
@@ -62,7 +62,7 @@ namespace Mangos.World.Player
 
 		private int gainStat(int level, double a3, double a2, double a1, double a0)
 		{
-			return checked((int)Math.Round(a3 * (double)level * (double)level * (double)level + a2 * (double)level * (double)level + a1 * (double)level + a0) - (int)Math.Round(a3 * (double)(level - 1) * (double)(level - 1) * (double)(level - 1) + a2 * (double)(level - 1) * (double)(level - 1) + a1 * (double)(level - 1) + a0));
+			return checked((int)Math.Round(a3 * level * level * level + a2 * level * level + a1 * level + a0) - (int)Math.Round(a3 * (level - 1) * (level - 1) * (level - 1) + a2 * (level - 1) * (level - 1) + a1 * (level - 1) + a0));
 		}
 
 		public void CalculateOnLevelUP(ref WS_PlayerData.CharacterObject objCharacter)
@@ -86,7 +86,7 @@ namespace Mangos.World.Player
 					}
 					if (objCharacter.Level <= 25)
 					{
-						objCharacter.Mana.Base += 20 + unchecked((int)objCharacter.Level);
+						objCharacter.Mana.Base += 20 + unchecked(objCharacter.Level);
 					}
 					else
 					{
@@ -105,11 +105,11 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) + 4;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) + 4;
 					}
 					if (objCharacter.Level <= 27)
 					{
-						objCharacter.Mana.Base += 18 + unchecked((int)objCharacter.Level);
+						objCharacter.Mana.Base += 18 + unchecked(objCharacter.Level);
 					}
 					else
 					{
@@ -128,11 +128,11 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) - 8;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) - 8;
 					}
 					if (objCharacter.Level <= 27)
 					{
-						objCharacter.Mana.Base += 23 + unchecked((int)objCharacter.Level);
+						objCharacter.Mana.Base += 23 + unchecked(objCharacter.Level);
 					}
 					else
 					{
@@ -151,11 +151,11 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) + 4;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) + 4;
 					}
 					if (objCharacter.Level <= 25)
 					{
-						objCharacter.Mana.Base += 17 + unchecked((int)objCharacter.Level);
+						objCharacter.Mana.Base += 17 + unchecked(objCharacter.Level);
 					}
 					else
 					{
@@ -174,11 +174,11 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) - 6;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) - 6;
 					}
 					if (objCharacter.Level <= 33)
 					{
-						objCharacter.Mana.Base += 22 + unchecked((int)objCharacter.Level);
+						objCharacter.Mana.Base += 22 + unchecked(objCharacter.Level);
 					}
 					else
 					{
@@ -201,7 +201,7 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) + 2;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) + 2;
 					}
 					objCharacter.Strength.Base += gainStat(objCharacter.Level, 2.5E-05, 0.00417, 0.654096, -0.601491);
 					objCharacter.Intellect.Base += gainStat(objCharacter.Level, 8E-06, 0.001001, 0.16319, -0.06428);
@@ -216,11 +216,11 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) + 1;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) + 1;
 					}
 					if (objCharacter.Level <= 32)
 					{
-						objCharacter.Mana.Base += 19 + unchecked((int)objCharacter.Level);
+						objCharacter.Mana.Base += 19 + unchecked(objCharacter.Level);
 					}
 					else
 					{
@@ -239,11 +239,11 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) - 2;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) - 2;
 					}
 					if (objCharacter.Level <= 30)
 					{
-						objCharacter.Mana.Base += 21 + unchecked((int)objCharacter.Level);
+						objCharacter.Mana.Base += 21 + unchecked(objCharacter.Level);
 					}
 					else
 					{
@@ -262,7 +262,7 @@ namespace Mangos.World.Player
 					}
 					else
 					{
-						objCharacter.Life.Base += unchecked((int)objCharacter.Level) + 10;
+						objCharacter.Life.Base += unchecked(objCharacter.Level) + 10;
 					}
 					objCharacter.Strength.Base += gainStat(objCharacter.Level, 3.9E-05, 0.006902, 1.0800399999999999, -1.051701);
 					objCharacter.Intellect.Base += gainStat(objCharacter.Level, 2E-06, 0.001003, 0.10089, -0.076055);
@@ -351,13 +351,13 @@ namespace Mangos.World.Player
 							}
 							j = (byte)unchecked((uint)(j + 1));
 						}
-						while (unchecked((uint)j) <= 3u);
+						while (unchecked(j) <= 3u);
 						break;
 					}
 					i = (byte)unchecked((uint)(i + 1));
 				}
 			}
-			while ((uint)i <= 63u);
+			while (i <= 63u);
 		}
 	}
 }
